게시물 삭제시 본인이 작성한 게시물, (로그인 아이디와 게시물객체의 로그인아이디가 일치할경우) 에만 삭제 처리 자신이 작성한 게시물에 대해서만 x버튼 보이도록 처리

+++ src/main/java/com/example/winku/controller/feed/FeedController.java
@@ -74,9 +74,12 @@ public class FeedController {
     }
 
     @PostMapping("/feed/deleteFeed")
-    public String feedDelete(@ModelAttribute DeleteFeedDto feedDto) {
+    public String feedDelete(@ModelAttribute DeleteFeedDto feedDto, @SessionAttribute(name = "user") User user, @RequestParam String redirect) {
+        if (!feedService.findFeedId(feedDto.getId()).getLoginId().equals(user.getLoginId())) {
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN);
+        }
         feedService.deleteFeed(feedDto);
-        return "redirect:/feed/index";
+        return "redirect:" + redirect;
     }
 
     @PostMapping("/feed/deleteComment")
